fixing shitty problems

+++ main.py
@@ -223,13 +223,12 @@ def delete_all_zip():
             #print("clutter cleared :" ,f)
 
 if __name__ == "__main__":
-    if not os.path.exists('temp_files'):
-        os.makedirs('temp_files')
     #this cycle happens when the app starts a new dyno
     delete_all_zip()    #for clearing up the clutter
     #delete the files in the folder 'temp_files' 
     list_of_files = os.listdir('temp_files')
     for f in list_of_files:
+        if f != 'test.txt':
             os.remove("temp_files/"+f)
             #print("clutter cleared from temp_files:" ,f)
     app.run(debug = False)
